fixed error messages
