--- src/_tests_/components/mock.json
@@ -1,139 +1,0 @@
-{
-    "documentation": "https://opencagedata.com/api",
-    "licenses": [
-        {
-            "name": "see attribution guide",
-            "url": "https://opencagedata.com/credits"
-        }
-    ],
-    "rate": {
-        "limit": 2500,
-        "remaining": 2498,
-        "reset": 1623715200
-    },
-    "results": [
-        {
-            "annotations": {
-                "DMS": {
-                    "lat": "51\u00b0 57' 9.57564'' N",
-                    "lng": "7\u00b0 37' 56.90280'' E"
-                },
-                "MGRS": "32UMC0602156656",
-                "Maidenhead": "JO31tw58vp",
-                "Mercator": {
-                    "x": 849643.007,
-                    "y": 6757899.137
-                },
-                "OSM": {
-                    "edit_url": "https://www.openstreetmap.org/edit?way=125838041#map=16/51.95266/7.63247",
-                    "note_url": "https://www.openstreetmap.org/note/new#map=16/51.95266/7.63247&layers=N",
-                    "url": "https://www.openstreetmap.org/?mlat=51.95266&mlon=7.63247#map=16/51.95266/7.63247"
-                },
-                "UN_M49": {
-                    "regions": {
-                        "DE": "276",
-                        "EUROPE": "150",
-                        "WESTERN_EUROPE": "155",
-                        "WORLD": "001"
-                    },
-                    "statistical_groupings": ["MEDC"]
-                },
-                "callingcode": 49,
-                "currency": {
-                    "alternate_symbols": [],
-                    "decimal_mark": ",",
-                    "html_entity": "&#x20AC;",
-                    "iso_code": "EUR",
-                    "iso_numeric": "978",
-                    "name": "Euro",
-                    "smallest_denomination": 1,
-                    "subunit": "Cent",
-                    "subunit_to_unit": 100,
-                    "symbol": "\u20ac",
-                    "symbol_first": 0,
-                    "thousands_separator": "."
-                },
-                "flag": "\ud83c\udde9\ud83c\uddea",
-                "geohash": "u1jrt9ty1t8rg3r5wttm",
-                "qibla": 128.55,
-                "roadinfo": {
-                    "drive_on": "right",
-                    "road": "Friedrich-Ebert-Stra\u00dfe",
-                    "speed_in": "km/h"
-                },
-                "sun": {
-                    "rise": {
-                        "apparent": 1623640260,
-                        "astronomical": 0,
-                        "civil": 1623637320,
-                        "nautical": 1623632760
-                    },
-                    "set": {
-                        "apparent": 1623700140,
-                        "astronomical": 0,
-                        "civil": 1623703080,
-                        "nautical": 1623707700
-                    }
-                },
-                "timezone": {
-                    "name": "Europe/Berlin",
-                    "now_in_dst": 1,
-                    "offset_sec": 7200,
-                    "offset_string": "+0200",
-                    "short_name": "CEST"
-                },
-                "what3words": {
-                    "words": "episodes.mass.below"
-                }
-            },
-            "bounds": {
-                "northeast": {
-                    "lat": 51.9528202,
-                    "lng": 7.6325938
-                },
-                "southwest": {
-                    "lat": 51.9525445,
-                    "lng": 7.6323594
-                }
-            },
-            "components": {
-                "ISO_3166-1_alpha-2": "DE",
-                "ISO_3166-1_alpha-3": "DEU",
-                "_category": "building",
-                "_type": "building",
-                "city": "M\u00fcnster",
-                "city_district": "M\u00fcnster-Mitte",
-                "continent": "Europe",
-                "country": "Alemanha",
-                "country_code": "de",
-                "house_number": "7",
-                "neighbourhood": "Josef",
-                "political_union": "European Union",
-                "postcode": "48153",
-                "road": "Friedrich-Ebert-Stra\u00dfe",
-                "state": "Ren\u00e2nia do Norte-Vestf\u00e1lia",
-                "suburb": "Innenstadtring"
-            },
-            "confidence": 10,
-            "formatted": "Friedrich-Ebert-Stra\u00dfe 7, 48153 M\u00fcnster, Alemanha",
-            "geometry": {
-                "lat": 51.9526599,
-                "lng": 7.632473
-            }
-        }
-    ],
-    "status": {
-        "code": 200,
-        "message": "OK"
-    },
-    "stay_informed": {
-        "blog": "https://blog.opencagedata.com",
-        "twitter": "https://twitter.com/OpenCage"
-    },
-    "thanks": "For using an OpenCage API",
-    "timestamp": {
-        "created_http": "Mon, 14 Jun 2021 00:47:21 GMT",
-        "created_unix": 1623631641
-    },
-    "total_results": 1
-}
