+++ src/services/GeolocationService.ts
@@ -1,6 +1,6 @@
 import { Coordinates } from '../interfaces/Coordinates';
 
-export const URL_GEOLOCATION =
+const URL_GEOLOCATION =
     'https://api.opencagedata.com/geocode/v1/json?key=e4a1d306c8bb4fdda513d15283580c71&lang=pt_br&limit=1';
 
 export const getCoordinatesByLocation = async (
@@ -8,7 +8,6 @@ export const getCoordinatesByLocation = async (
 ): Promise<Coordinates> => {
     const response = await fetch(`${URL_GEOLOCATION}&q=${location}`);
     const data = await response.json();
-
     if (data.results?.length < 1) {
         return Promise.reject();
     }
